remove-lines: Parse code, only remove lines from whitespace
Fixes #7

removeLines currently removes blank lines from source code
without parsing the source,
which leads to it removing blank lines from template strings.

This change uses acorn to parse the code,
which is already used to parse code in blank-comments,
and provides an `onToken` handler.
This handler gets called for every token acorn emits.
acorn does not emit tokens for whitespace,
so our handler can find whitespace by storing a position,
and advancing it to the end of each emitted token;
every gap between the position and the start of the next token
is guaranteed to be a region of whitespace.

acorn does generate tokens for template strings,
preventing our code from accidentally modifying them.

Squashing a region of whitespace works like the old code did.
Since the region may not be at the beginning or the end of the code,
we add some extra checks for those conditions
and handle initial blank lines or trailing whitespace
only if they are met.

Other changes of note:

* Since we only handle whitespace regions,
  we had to modify FIRST_LINES to work in them.

* There is no good place to increment maxEolChars,
  so instead we generate both values early
  and pass the appropriate one to limitLines as a parameter.